Sync changes from hyperweb-boilerplate for commit 4359eec
