--- templates/hyperweb/__tests__/first.test.ts
@@ -1,5 +1,0 @@
-import { join } from 'path';
-
-it('works', () => {
-    console.log('hello test world!' + join('a/d/d/', 'b/c'));
-})
