+++ templates/hyperweb/configs/ci.yaml
@@ -13,14 +13,15 @@ chains:
       faucet: 8000
       exposer: 8001
     resources:
-      cpu: "0.3"
-      memory: "300M"
+      cpu: "0.4"
+      memory: "400M"
     faucet:
       enabled: true
       type: starship
+      concurrency: 2
       resources:
-        cpu: "0.1"
-        memory: "100M"
+        cpu: "0.2"
+        memory: "200M"
 
 registry:
   enabled: true
