+++ templates/hyperweb/package.json
@@ -39,8 +39,8 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-plugin-unused-imports": "^3.0.0",
-    "hyperwebjs": "0.0.4",
     "jest": "^29.6.2",
+    "hyperwebjs": "0.0.4",
     "prettier": "^3.0.2",
     "rimraf": "4.4.1",
     "starshipjs": "^2.4.1",
