+++ templates/hyperweb/src/amm-contract/index.ts
